hook for watching class
