+++ frontend/components/ResultsPage/Results/MobileSearchResult.tsx
@@ -4,11 +4,11 @@ import Course from '../../classModels/Course'
 import IconCollapseExpand from '../../images/IconCollapseExpand'
 import IconArrow from '../../images/IconArrow'
 import useResultRequisite from './useResultRequisite'
+import useUserChange from './useUserChange'
 import macros from '../../macros'
 import MobileSectionPanel from './MobileSectionPanel'
 import SignUpForNotifications from '../../SignUpForNotifications'
-import user from '../../user'
-import Keys from '../../../../common/Keys'
+
 
 
 
@@ -23,7 +23,7 @@ function MobileSearchResult({ aClass, history } : MobileSearchResultProps) {
   const [showNUPath, setShowNUPath] = useState(false)
   const [showPrereq, setShowPrereq] = useState(false)
   const [showCoreq, setShowCoreq] = useState(false)
-  const [userIsWatchingClass, setUserIsWatchingClass] = useState(user.isWatchingClass(Keys.getClassHash(aClass)))
+  const userIsWatchingClass = useUserChange(aClass)
   const [showAll, setShowAll] = useState(false)
   
   const sectionsShownByDefault = aClass.sections.length < 3 ? aClass.sections.length : 3
@@ -31,19 +31,6 @@ function MobileSearchResult({ aClass, history } : MobileSearchResultProps) {
   const hideShowAll =  sectionsShownByDefault === aClass.sections.length
   const optionalDisplay = useResultRequisite(history);
 
-
-  const onUserUpdate = () => {
-    // Show the notification toggles if the user is watching this class.
-    const isWatching = user.isWatchingClass(Keys.getClassHash(aClass));
-    if (isWatching !== userIsWatchingClass) {
-      setUserIsWatchingClass(isWatching)
-    }
-  }
-
-  useEffect(() => {
-    user.registerUserChangeHandler(onUserUpdate)
-    return () => user.unregisterUserChangeHandler(onUserUpdate)
-  }, [])
 
   useEffect(() => {
     if (showAll) {
